Updated based on review comments
1. Extended both iterators from a common profile iterator
2. Other miscellaneous comments
