+++ exec/java-exec/src/main/java/org/apache/drill/exec/ops/FragmentContext.java
@@ -230,8 +230,8 @@ public class FragmentContext extends BaseFragmentContext implements AutoCloseabl
     return context;
   }
 
-  public QueryContext getQueryContext() {
-    return queryContext;
+  public boolean isUserAuthenticationEnabled() {
+    return queryContext.isUserAuthenticationEnabled();
   }
 
   /**
