+++ exec/java-exec/src/main/java/org/apache/drill/exec/store/sys/ProfileJsonIterator.java
@@ -17,112 +17,66 @@
  */
 package org.apache.drill.exec.store.sys;
 
-import com.google.common.base.Function;
-import com.google.common.collect.Iterators;
-
-import org.apache.drill.exec.ExecConstants;
-import org.apache.drill.exec.ops.FragmentContext;
-import org.apache.drill.exec.ops.QueryContext;
-import org.apache.drill.exec.proto.UserBitShared;
-import org.apache.drill.exec.proto.UserBitShared.QueryProfile;
-import org.apache.drill.exec.serialization.InstanceSerializer;
-import org.apache.drill.exec.server.QueryProfileStoreContext;
-import org.apache.drill.exec.server.options.QueryOptionManager;
-import org.apache.drill.exec.util.ImpersonationUtil;
-
-import javax.annotation.Nullable;
 import java.io.IOException;
 import java.util.Iterator;
-import java.util.LinkedList;
-import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 
+import javax.annotation.Nullable;
+
+import org.apache.drill.exec.ops.FragmentContext;
+import org.apache.drill.exec.proto.UserBitShared;
+import org.apache.drill.exec.proto.UserBitShared.QueryProfile;
+import org.apache.drill.exec.serialization.InstanceSerializer;
+
+import com.google.common.base.Function;
+import com.google.common.collect.Iterators;
+
 /**
- * DRILL-5068: Add a new system table for completed profiles (JSON)
+ * System table listing completed profiles as JSON documents
  */
-public class ProfileJsonIterator implements Iterator<Object> {
+public class ProfileJsonIterator extends ProfileIterator {
   static final org.slf4j.Logger logger = org.slf4j.LoggerFactory.getLogger(ProfileJsonIterator.class);
 
-  private final QueryProfileStoreContext profileStoreContext;
   private final InstanceSerializer<QueryProfile> profileSerializer;
-  private final Iterator<ProfileInfoJson> itr;
-  private final String queryingUsername;
-  private final boolean isAdmin;
+  private final Iterator<ProfileJson> itr;
 
   public ProfileJsonIterator(FragmentContext context) {
-    //Grab profile Store Context
-    profileStoreContext = context
-        .getDrillbitContext()
-        .getProfileStoreContext();
-
+    super(context);
     //Holding a serializer (for JSON extract)
     profileSerializer = profileStoreContext.
         getProfileStoreConfig().getSerializer();
 
-    queryingUsername = context.getQueryUserName();
-    isAdmin = hasAdminPrivileges(context.getQueryContext());
-
-    itr = iterateProfileInfoJson(context);
-  }
-
-  private boolean hasAdminPrivileges(QueryContext queryContext) {
-    if (queryContext == null) {
-      return false;
-    }
-    QueryOptionManager options = queryContext.getOptions();
-    if (queryContext.isUserAuthenticationEnabled() &&
-        !ImpersonationUtil.hasAdminPrivileges(
-          queryContext.getQueryUserName(),
-          ExecConstants.ADMIN_USERS_VALIDATOR.getAdminUsers(options),
-          ExecConstants.ADMIN_USER_GROUPS_VALIDATOR.getAdminUserGroups(options))) {
-      return false;
-    }
-
-    //Passed checks
-    return true;
-  }
-
-  private Iterator<ProfileInfoJson> iterateProfileInfoJson(FragmentContext context) {
-    try {
-      PersistentStore<UserBitShared.QueryProfile> profiles = profileStoreContext.getCompletedProfileStore();
-
-      return transformJson(getAuthorizedProfiles(profiles.getAll(), queryingUsername, isAdmin));
-
-    } catch (Exception e) {
-      logger.error(String.format("Unable to get persistence store: %s, ", profileStoreContext.getProfileStoreConfig().getName(), e));
-      return Iterators.singletonIterator(ProfileInfoJson.getDefault());
-    }
+    itr = iterateProfileInfoJson();
   }
 
   //Returns an iterator for authorized profiles
-  private Iterator<Entry<String, QueryProfile>> getAuthorizedProfiles (
-      Iterator<Entry<String, QueryProfile>> allProfiles, String username, boolean isAdministrator) {
-    if (isAdministrator) {
-      return allProfiles;
-    }
+  private Iterator<ProfileJson> iterateProfileInfoJson() {
+    try {
+      //Transform authorized profiles to iterator for ProfileInfoJson
+      return transformJson(
+          getAuthorizedProfiles(
+            profileStoreContext
+              .getCompletedProfileStore()
+              .getAll(),
+            queryingUsername, isAdmin));
 
-    List<Entry<String, QueryProfile>> authorizedProfiles = new LinkedList<Entry<String, QueryProfile>>();
-    while (allProfiles.hasNext()) {
-      Entry<String, QueryProfile> profileKVPair = allProfiles.next();
-      //Check if user matches
-      if (profileKVPair.getValue().getUser().equals(username)) {
-        authorizedProfiles.add(profileKVPair);
-      }
+    } catch (Exception e) {
+      logger.error(e.getMessage());
+      return Iterators.singletonIterator(ProfileJson.getDefault());
     }
-    return authorizedProfiles.iterator();
   }
 
   /**
-   * Iterating persistentStore as a iterator of {@link org.apache.drill.exec.store.sys.ProfileJsonIterator.ProfileInfoJson}.
+   * Iterating persistentStore as a iterator of {@link org.apache.drill.exec.store.sys.ProfileJsonIterator.ProfileJson}.
    */
-  private Iterator<ProfileInfoJson> transformJson(Iterator<Map.Entry<String, UserBitShared.QueryProfile>> all) {
-    return Iterators.transform(all, new Function<Map.Entry<String, UserBitShared.QueryProfile>, ProfileInfoJson>() {
+  private Iterator<ProfileJson> transformJson(Iterator<Map.Entry<String, UserBitShared.QueryProfile>> all) {
+    return Iterators.transform(all, new Function<Map.Entry<String, UserBitShared.QueryProfile>, ProfileJson>() {
       @Nullable
       @Override
-      public ProfileInfoJson apply(@Nullable Map.Entry<String, UserBitShared.QueryProfile> input) {
+      public ProfileJson apply(@Nullable Map.Entry<String, UserBitShared.QueryProfile> input) {
         if (input == null || input.getValue() == null) {
-          return ProfileInfoJson.getDefault();
+          return ProfileJson.getDefault();
         }
 
         //Constructing ProfileInfo
@@ -135,7 +89,7 @@ public class ProfileJsonIterator implements Iterator<Object> {
           profileJson = "{ 'message' : 'error (unable to serialize profile: "+ queryID +")' }";
         }
 
-        return new ProfileInfoJson(
+        return new ProfileJson(
             queryID,
             profileJson
          );
@@ -158,20 +112,20 @@ public class ProfileJsonIterator implements Iterator<Object> {
     throw new UnsupportedOperationException();
   }
 
-  public static class ProfileInfoJson {
-    private static final String UnknownValue = "UNKNOWN";
+  public static class ProfileJson {
+    private static final String UnknownValue = "N/A";
 
-    private static final ProfileInfoJson DEFAULT = new ProfileInfoJson();
+    private static final ProfileJson DEFAULT = new ProfileJson();
 
     public final String queryId;
     public final String json;
 
-    public ProfileInfoJson(String query_id, String profileJson) {
+    public ProfileJson(String query_id, String profileJson) {
       this.queryId = query_id;
       this.json = profileJson;
     }
 
-    private ProfileInfoJson() {
+    private ProfileJson() {
       this(UnknownValue, UnknownValue);
     }
 
@@ -179,10 +133,8 @@ public class ProfileJsonIterator implements Iterator<Object> {
      * If unable to get ProfileInfo, use this default instance instead.
      * @return the default instance
      */
-    public static final ProfileInfoJson getDefault() {
+    public static final ProfileJson getDefault() {
       return DEFAULT;
     }
   }
 }
-
-
